# Conflicts:
#	.idea/libraries/Maven__io_netty_netty_codec_http2_4_1_43_Final.xml
#	.idea/libraries/Maven__io_netty_netty_codec_http_4_1_43_Final.xml
#	.idea/libraries/Maven__io_netty_netty_codec_socks_4_1_43_Final.xml
#	.idea/libraries/Maven__io_netty_netty_handler_proxy_4_1_43_Final.xml
#	.idea/libraries/Maven__io_netty_netty_transport_native_epoll_linux_x86_64_4_1_43_Final.xml
#	.idea/libraries/Maven__io_netty_netty_transport_native_unix_common_4_1_43_Final.xml
#	.idea/libraries/Maven__io_projectreactor_netty_reactor_netty_0_9_2_RELEASE.xml
#	.idea/libraries/Maven__org_glassfish_jakarta_el_3_0_3.xml
#	.idea/libraries/Maven__org_springframework_boot_spring_boot_starter_reactor_netty_2_2_2_RELEASE.xml
#	.idea/libraries/Maven__org_springframework_boot_spring_boot_starter_webflux_2_2_2_RELEASE.xml
#	.idea/libraries/Maven__org_springframework_spring_webflux_5_2_2_RELEASE.xml
#	.idea/libraries/Maven__org_synchronoss_cloud_nio_multipart_parser_1_1_0.xml
#	.idea/libraries/Maven__org_synchronoss_cloud_nio_stream_storage_1_1_3.xml
#	sosotaxi/driverState-service/driverState-service.iml
#	sosotaxi/pom.xml
#	sosotaxi/user-service/user-service.iml

+++ sosotaxi/dispatch-service/src/main/java/com/apicaller/sosotaxi/service/impl/DispatchServiceImpl.java
@@ -1,6 +1,7 @@
 package com.apicaller.sosotaxi.service.impl;
 
 import com.alibaba.fastjson.JSONObject;
+import com.apicaller.sosotaxi.entity.MinimizedDriver;
 import com.apicaller.sosotaxi.entity.UnsettledOrder;
 import com.apicaller.sosotaxi.entity.GeoPoint;
 import com.apicaller.sosotaxi.service.DispatchService;
@@ -9,11 +10,86 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Resource;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 @Component
 public class DispatchServiceImpl implements DispatchService {
+
+    /**
+     * 派单模式
+     */
+    public enum dispatchMethod{
+        /**
+         * 抢单模式
+         */
+        GROUPDISPATCH,
+        /**
+         * 自动安排模式
+         */
+        AUTOARRANGE
+    }
+
+    /**
+     * 当前策略
+     */
+    private dispatchMethod currentMethod = dispatchMethod.AUTOARRANGE;
+
+    /**
+     * 更改策略
+     * @param method
+     * @return
+     */
+    public Boolean changeDispatchMethod(String method){
+        if(method.equals(dispatchMethod.AUTOARRANGE.toString())){
+            currentMethod = dispatchMethod.AUTOARRANGE;
+        }
+        else if(method.equals(dispatchMethod.GROUPDISPATCH.toString())){
+            currentMethod = dispatchMethod.GROUPDISPATCH;
+        }
+        else{
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * 以字符串的形式返回当前策略
+     * 后期分配策略的部分稳定后可以换成返回枚举，因为目前还要调整所以暂时用字符串减少依赖
+     * @return
+     */
+    public String getDispatchedMethod(){
+        if(currentMethod == dispatchMethod.AUTOARRANGE){
+            return "singleDispatch";
+        }
+        else if(currentMethod == dispatchMethod.GROUPDISPATCH){
+            return "groupDispatch";
+        }
+        else{
+            return "still developing ...";
+        }
+    }
+
+    /**
+     * 抢单模式下一单的最大分配司机数
+     */
+    static private int arrangeCount = 8;
+
+    /**
+     * 设置最大分配司机数
+     * 仅当模式为抢单模式时有效
+     * @param arrangeCount
+     */
+    public void setArrangeCount(int arrangeCount){
+        DispatchServiceImpl.arrangeCount = arrangeCount;
+    }
+
+    /**
+     * 设置最大分配司机数
+     * @return
+     */
+    public int getArrangeCount(){
+        return DispatchServiceImpl.arrangeCount;
+    }
 
     @Resource
     private InfoCacheServiceImpl infoCacheService;
@@ -25,6 +101,10 @@ public class DispatchServiceImpl implements DispatchService {
     @Override
     @Scheduled(fixedDelay = 5000)
     public void dispatch() throws Exception {
+        //仅当当前策略为抢单模式时才执行定时处理方法
+        if(currentMethod!=dispatchMethod.GROUPDISPATCH){
+            return;
+        }
         List<UnsettledOrder> orders = infoCacheService.getAllUOrder();
         Set<String> driverIds = infoCacheService.getAllDrivers();
         infoCacheService.deleteDurationSets(infoCacheService.getAllDurationSetKeys());
@@ -52,7 +132,7 @@ public class DispatchServiceImpl implements DispatchService {
             }
             Set<String> driversToDispatch = null;
             try {
-                driversToDispatch = infoCacheService.getDriverIdByRank(order.getOrderId());
+                driversToDispatch = infoCacheService.getDriverIdByRank(order.getOrderId(), arrangeCount);
             }catch (Exception e){
                 e.printStackTrace();
             }
@@ -64,6 +144,99 @@ public class DispatchServiceImpl implements DispatchService {
             for(String driverId:driversToDispatch){
                 infoCacheService.updateDispatch(driverId, order);
             }
+        }
+    }
+
+
+    /**
+     * 立即获取相应订单最合适的司机
+     * @param order
+     * @return 分配的司机
+     * 如果为空代表当前无合适司机，请等待一段时间后重试
+     * @throws Exception
+     */
+    @Override
+    public MinimizedDriver dispatch(UnsettledOrder order) throws Exception {
+        if(order == null){
+            throw new Exception("order不存在，传入了错误的参数");
+        }
+        String city = order.getCity();
+        int type = order.getType();
+        Set<String> drivers = infoCacheService.getHashByPattern(city, type);
+        if(drivers == null){
+            return null;
+        }
+        HashMap<String, Integer> driversMap = new HashMap<>();
+//        String optimalDriver = null;
+//        Integer minTime = Integer.MAX_VALUE;
+        for(String driverId:drivers){
+            GeoPoint point = null;
+            try{
+                point = infoCacheService.getDriverPosition(driverId);
+            }catch (Exception e){
+                e.printStackTrace();
+                throw new Exception("缓存中不存在该类型司机信息");
+            }
+            String result = null;
+            try{
+                if(point != null){
+                    result = mapUtils.getDirection(point, order.getOriginPoint());
+                }
+                else{
+                    continue;
+                }
+            }catch (Exception e){
+                e.printStackTrace();
+                throw new Exception("在分配算法中调用地图api时出错");
+            }
+            Integer durection = null;
+            try{
+                if(result != null){
+                    durection = MapUtils.getDistance(result);
+                }
+                else{
+                    continue;
+                }
+            }catch (Exception e){
+                e.printStackTrace();
+                throw new Exception("在分配算法中解析api返回结果时出错");
+            }
+            driversMap.put(driverId, durection);
+        }
+        if(driversMap.isEmpty()){
+            return null;
+        }
+        else{
+            //排序
+            List<Map.Entry<String, Integer>> list = new ArrayList<Map.Entry<String, Integer>>(driversMap.entrySet());
+            Collections.sort(list, new Comparator<Map.Entry<String, Integer>>() {
+                @Override
+                public int compare(Map.Entry<String, Integer> o1, Map.Entry<String, Integer> o2) {
+                    return o1.getValue().compareTo(o2.getValue());
+                }
+            });
+
+            MinimizedDriver targetDriver = null;
+
+            String orderId = order.getOrderId();
+
+            //遍历排序过的集合
+            for(Map.Entry<String, Integer> entry:list){
+                String driverId = entry.getKey();
+                //已经尝试过分配给该司机，不应该再次尝试
+                if(infoCacheService.getDispatchedSet(orderId) != null && infoCacheService.getDispatchedSet(orderId).contains(driverId)){
+                    continue;
+                }
+                //分配成功
+                //dispatchDriver方法会自动注销该司机，所以先拿出来
+                targetDriver = infoCacheService.getDriver(driverId);
+                if(infoCacheService.dispatchDriver(driverId)){
+                    //记录该司机已经尝试过接单
+                    infoCacheService.addDriverToDispatchedSet(orderId,driverId);
+                    break;
+                }
+            }
+            return targetDriver;
         }
     }
 }
